# bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WK API",
    "description": "We Kare Api 1.0 ",
    "termsOfService": "http://www.wkya.rw",
    "contact": {
      "name": "wecare",
      "url": "http://www.wkya.rw",
      "email": "rwahamanick@gmail.com"
    },
    "version": "v1",
    "x-logo": {
      "url": "/img/tiger-soft-logo.png",
      "backgroundColor": "#FFFFFF",
      "altText": "PetStore Logo"
    }
  },
  "paths": {
...
    "/barbers/partner/{partnerId}": {
      "get": {
        "tags": [
          "Barber"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "partnerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/businessServices": {
      "get": {
        "tags": [
          "BusinessService"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "BusinessService"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReferenceInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/businessServices/{id}": {
      "get": {
        "tags": [
          "BusinessService"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "BusinessService"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReferenceInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "BusinessService"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/partners": {
      "get": {
        "tags": [
          "Partner"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Partner"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PartnerInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/partners/{id}": {
      "get": {
        "tags": [
          "Partner"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Partner"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PartnerInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Partner"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/partners/updateLocation": {
      "put": {
        "tags": [
          "Partner"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PartnerLocationInputDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/partnerServices/partner/{partnerId}": {
      "get": {
        "tags": [
          "PartnerService"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "partnerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/partnerServices/service/{serviceId}": {
      "get": {
        "tags": [
          "PartnerService"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "serviceId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/reservations/partner/{partnerId}": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "partnerId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ReferenceInputDto": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "PartnerInputDto": {
      "type": "object",
      "properties": {
        "name": {
          "maxLength": 250,
          "type": "string"
        },
        "location": {
          "maxLength": 300,
          "type": "string"
        },
        "email": {
          "maxLength": 125,
          "type": "string"
        },
        "phone": {
          "maxLength": 10,
          "type": "string"
        },
        "momoCode": {
          "maxLength": 8,
          "type": "string"
        },
        "userId": {
          "format": "uuid",
          "type": "string"
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "PartnerLocationInputDto": {
      "type": "object",
      "properties": {
        "partnerId": {
          "format": "int32",
          "type": "integer"
        },
        "destinationLatitude": {
          "type": "string"
        },
        "destinationLongitude": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Authorization"
    }
  },
  "security": [
    { }
  ]
}